Organize components by-feature insteas of by-type
Changes:
- move AppCmp to app/core
- move HomeCmp to app/home
- move AboutCmp to app/about
- move NameList service to shared/services
- move AppCmp to app/app (As per the suggestions in PR#382)
- merge [feature]/tests into [feature]/components (As per the suggestions in PR#382)
- rename bootstrap.ts/hot_loader_bootstrap.tx -> main.ts/hot_loader_main.ts
- update README.md to reflect the new structure
- remote unused call to HTTP_PROVIDERS in app.ts
- fix link in tools/tasks/build.bundles.ts - bootstrap -> main
