+++ app/about/components/about_spec.ts
@@ -8,7 +8,7 @@ import {
 import {Component} from 'angular2/core';
 import {DOM} from 'angular2/src/platform/dom/dom_adapter';
 import {AboutCmp} from './about';
-import {NameList} from '../../services/name_list';
+import {NameList} from '../../shared/services/name_list';
 
 export function main() {
   describe('About component', () => {
