+++ app/home/components/home.ts
@@ -2,7 +2,7 @@ import {Component} from 'angular2/core';
 
 @Component({
   selector: 'home',
-  templateUrl: './components/home/home.html',
-  styleUrls: ['./components/home/home.css']
+  templateUrl: './home/components/home.html',
+  styleUrls: ['./home/components/home.css']
 })
 export class HomeCmp {}
