+++ app/hot_loader_main.ts
@@ -2,7 +2,7 @@
 
 import {provide} from 'angular2/core';
 import {ROUTER_PROVIDERS, LocationStrategy, HashLocationStrategy} from 'angular2/router';
-import {AppCmp} from './components/app/app';
+import {AppCmp} from './app/components/app';
 
 System.import('//localhost:<%= HOT_LOADER_PORT %>/ng2-hot-loader')
   .then(loader => {
