+++ app/main.ts
@@ -1,7 +1,7 @@
 import {provide} from 'angular2/core';
 import {bootstrap} from 'angular2/platform/browser';
 import {ROUTER_PROVIDERS, LocationStrategy, HashLocationStrategy} from 'angular2/router';
-import {AppCmp} from './components/app/app';
+import {AppCmp} from './app/components/app';
 
 bootstrap(AppCmp, [
   ROUTER_PROVIDERS,
